Send moves when reconnect
This is a first attempt to record moves when a connection drops so they 
can be resent later.

This only records and resends all moves. It still looses moves if page 
refreshed (this could be done by using local storage rather than session 
storage, but I wanted to think about expiring before I did this) and it 
does not handle conflicts well (last to send always wins).