Try to make re-syncing more robust
I noticed that re-syncing didn't work and caused some letters to be lost 
when I had a poor connection at the point of trying to sync. Looking at 
the actioncable source, I can see the call to `perform` ultimatly 
returns the result of `send`[1].

`send` returns and sends the payload based on if the socket `isOpen` 
which is different to `isActive` so may account for the problem if the 
socket was in the `connecting` state.

[1] 
https://github.com/rails/rails/blob/c96139af71e6f7c36e25bccea6b05ccd9523531a/actioncable/app/javascript/action_cable/connection.js#L20